Reorganize db specific classifiers
