+++ nlp/__init__.py
@@ -1,6 +1,6 @@
 from nlp.Parser import Parser
 from nlp.Tagger import Tagger
 from nlp.Tokenizer import Tokenizer
-from nlp.SimFinder import SimFinder
-from nlp.CorpusClassifier import CorpusClassifier
-from nlp.CorpusGenerator import CorpusGenerator
+from nlp.DBSchemaClassifier import DBSchemaClassifier
+from nlp.DBCorpusClassifier import DBCorpusClassifier
+from nlp.DBCorpusGenerator import DBCorpusGenerator
